Move plugin.cfg to root directory
I want to be able to add this plugin to my godot project using git submodules.

However, previously the plugin.cfg file was in a subdirectory. So if I cloned
the plugin code into addons/, it would not be recognised by godot, since it must
be under addons/<name>/plugin.cfg instead of addons/<name>/.../plugin.cfg

This commit moves plugin.cfg into the base directory of the repository, so that
it can be directly cloned into addons/ and be recognised by godot.

Also fixes .gitignore to work properly in a submodule.